docs:Chapter08 mission

+++ mino/springboot/main/java/org/example/swaggerpr/review/entity/Review.java
@@ -1,9 +1,20 @@
 package org.example.swaggerpr.review.entity;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.example.swaggerpr.member.entity.Member;
-import org.example.swaggerpr.mission.entity.Store;
+import org.example.swaggerpr.store.entity.Store;
 
 import java.time.LocalDateTime;
 
@@ -21,19 +32,16 @@ public class Review {
     @Column(nullable = false)
     private LocalDateTime createdAt;
 
-    // 현재 API 명세에는 텍스트 리뷰만 존재
     @Column(nullable = false, columnDefinition = "TEXT")
     private String content;
 
     @Column(nullable = false)
-    private Float score;
+    private Integer score;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
     private Member member;
 
-    // 리뷰 작성 API는 missionId를 받지만 ERD에서는 Review가 Store에 연결된다.
-    // 서비스에서 missionId로 Mission을 조회한 뒤 Mission.store를 저장하도록 설계했다.
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "store_id", nullable = false)
     private Store store;
